Restructure src directory and edited filepath imports.
